testing #8: Added the profile stuff for following.

+++ app/helpers/users_helper.rb
@@ -1,9 +1,13 @@
 module UsersHelper
-  def follow
-    if UserRelationship.where({:follower_id => current_user.id, :guide_id => params[:id]}).first.nil?
-      "#{button_to 'Follow', follow_author_path(params[:id]), :remote => true}"
+  def follow(guide_id)
+    if current_user.id != guide_id
+      if UserRelationship.where({:follower_id => current_user.id, :guide_id => params[:id]}).first.nil?
+        return "#{button_to 'Follow', follow_author_path(params[:id]), :remote => true}"
+      else
+        return "#{button_to 'Unfollow', unfollow_author_path(params[:id]), :remote => true, :method => :delete}"
+      end
     else
-      "#{button_to 'Unfollow', unfollow_author_path(params[:id]), :remote => true, :method => :delete}"
+      return ''
     end
   end
 end
